requirements.txt
requests
web3
python-dotenv
websocket-client